I made changes to the buttons paddings to make it uniform padding.
I also made some changes to the white navbar so as to make it responsive on the tablet screen size and above.

I added top margin and a button component to the "How Does It Work" section.

I also made some changes to the "Subscribe" section following the Figma design guidelines.

I added flex percentages to the Map and the Map Information section.

I changed some icons used from the react-icon library to the downloaded icons from the Figma design.