Created a restaurant-expanded-page with all the necessary information and components.
I made a change in the button in the navbar component so as to allow the use of black background in some components.

I gave the restaurant-expanded-page a route that would be taking an "id", so as to navigate between different restaurant page depending on the id that would be assigned.